fabfile: add local_shell task

+++ sandy-disaster-recovery/fabfile.py
@@ -1,5 +1,5 @@
 
-import os, shutil
+import os, shutil, json
 from glob import glob
 from tempfile import mkdtemp, gettempdir
 
@@ -13,32 +13,38 @@ import yaml
 
 # constants
 
+REQUIRED_SDK_VERSION = '1.8.8'
 APP_YAML_FILENAME = 'app.yaml'
 APP_YAML_TEMPLATE_FILENAME = 'app.yaml.template'
 BUILD_DIR_PREFIX = 'ccbuild'
 
 
-# find GAE appcfg
+# find GAE SDK
 
-GAE_APPCFG_POSSIBLE_LOCATIONS = [
-    '../../google_appengine/appcfg.py',
-    '../../../google_appengine/appcfg.py',
+POSSIBLE_SDK_DIRECTORIES = [
+    '../google_appengine/',
+    '../../google_appengine/',
+    '../../../google_appengine/',
 ]
 
 try:
-    _appcfg_path = (
-        path for path in GAE_APPCFG_POSSIBLE_LOCATIONS
+    _sdk_path = (
+        path for path in POSSIBLE_SDK_DIRECTORIES
         if os.path.exists(path)
+        and os.path.exists(os.path.join(path, 'appcfg.py'))
     ).next()
+    _sdk_path = os.path.join(*os.path.split(_sdk_path)[:-1])
 except StopIteration:
-    abort('appcfg.py not found - edit GAE_APPCFG_POSSIBLE_LOCATIONS')
+    abort('GAE SDK directory not found - add it to POSSIBLE_SDK_DIRECTORIES')
 
 
 # define env
 
 env.master_branch = "master"
 env.default_gae_app_version = "live"
-env.appcfg = os.path.realpath(_appcfg_path)
+env.sdk_path = os.path.realpath(_sdk_path)
+env.appcfg = os.path.realpath(os.path.join(_sdk_path, 'appcfg.py'))
+env.sdk_path = os.path.realpath(_sdk_path)
 
 
 # define apps
@@ -48,24 +54,38 @@ MINI_YAML = """
   secure: always
   allow_unclean_deploy: true
   sandbox: true
+  overwrite:
+    "assets/images/crisis-cleanup-logo-default.png" : "assets/images/crisis-cleanup-logo.png"
 
 - application: sandy-disaster-recovery
   secure: always
+  overwrite:
+    "assets/images/crisis-cleanup-logo-default.png" : "assets/images/crisis-cleanup-logo.png"
 
 - application: crisiscleanup-demo
   secure: optional
+  overwrite:
+    "assets/images/crisis-cleanup-logo-default.png" : "assets/images/crisis-cleanup-logo.png"
 
 - application: crisis-cleanup-au
   secure: always
+  overwrite:
+    "assets/images/crisis-cleanup-logo-au.png" : "assets/images/crisis-cleanup-logo.png"
 
 - application: crisis-cleanup-au-demo
   secure: optional
+  overwrite:
+    "assets/images/crisis-cleanup-logo-au.png" : "assets/images/crisis-cleanup-logo.png"
 
 - application: crisis-cleanup-in
   secure: always
+  overwrite:
+    "assets/images/crisis-cleanup-logo-in.png" : "assets/images/crisis-cleanup-logo.png"
 
 - application: crisis-cleanup-in-demo
   secure: optional
+  overwrite:
+    "assets/images/crisis-cleanup-logo-in.png" : "assets/images/crisis-cleanup-logo.png"
 """
 
 APPS = {
@@ -91,6 +111,19 @@ def warn_or_abort(app_defn, message):
         abort(message)
 
 
+def sdk_version_ok():
+    " Check that the SDK is the specified version. "
+    try:
+        sdk_version_d = yaml.load(open(os.path.join(env.sdk_path, 'VERSION')))
+        current_sdk_version = sdk_version_d['release']
+    except:
+        current_sdk_version = None
+    if current_sdk_version != REQUIRED_SDK_VERSION:
+        abort("Local SDK version is %s - %s is required" % (
+            current_sdk_version, REQUIRED_SDK_VERSION))
+    return True
+
+
 def app_yaml_template_present():
     " Check that the app.yaml template is present. "
     if APP_YAML_TEMPLATE_FILENAME not in os.listdir('.'):
@@ -99,13 +132,17 @@ def app_yaml_template_present():
 
 
 def working_directory_clean(app_defn):
-    " Check that the working directory is clean. "
+    """
+    Check that the working directory is clean.
+
+    Warn only; deployment is from a git commit.
+    """
     git_status = local("git status", capture=True)
     if "working directory clean" not in git_status:
         if app_defn.get('allow_unclean_deploy', False):
             warn("Working directory not clean (ignoring for %s)" % app_defn['application'])
         else:
-            abort("Working directory must be clean to deploy to %s" % app_defn['application'])
+            warn("Working directory not clean - modified files will not be deployed to %s" % app_defn['application'])
     return True
 
 
@@ -125,33 +162,52 @@ def on_master_branch(app_defn):
     return True
 
 
-def master_pushed_to_remote(app_defn):
-    " Check that the master branch is pushed to origin. "
-    local_master_ref = local("git rev-parse %s" % env.master_branch, capture=True)
-    origin_master_ref = local("git rev-parse origin/%s" % env.master_branch, capture=True)
-    if local_master_ref != origin_master_ref:
+def current_branch_pushed_to_remote(app_defn):
+    " Check that the current branch is pushed to origin. "
+    current_branch = get_current_branch()
+    local_ref = local("git rev-parse %s" % current_branch, capture=True)
+    origin_ref = local("git rev-parse origin/%s" % current_branch, capture=True)
+    if local_ref != origin_ref:
         if app_defn.get('allow_unclean_deploy', False):
             warn("%s and origin/%s differ (ignoring for %s)" % (
-                env.master_branch, env.master_branch, app_defn['application']))
+                current_branch, current_branch, app_defn['application']))
         else:
             abort("%s must be pushed to origin to deploy to %s" % (
-                env.master_branch, app_defn['application']))
+                current_branch, app_defn['application']))
     return True
 
 
-def ok_to_deploy(app_defn):
+def check_specified_commitish_pushed_to_remote(app_defn, tag):
+    " Check the specified commit on the current branch has been pushed to origin. "
+    current_branch = get_current_branch()
+    containing_branches = local("git branch -a --contains %s" % tag, capture=True)
+    if 'origin/%s' % current_branch not in containing_branches:
+        if app_defn.get('allow_unclean_deploy', False):
+            warn("%s has not been pushed to origin/%s (ignoring for %s)" % (
+                tag, current_branch, app_defn['application']))
+        else:
+            abort("%s must be pushed to origin/%s to deploy to %s" % (
+                tag, current_branch, app_defn['application']))
+
+    return True
+
+
+def ok_to_deploy(app_defn, tag):
     return (
+        sdk_version_ok() and
         app_yaml_template_present() and
         working_directory_clean(app_defn) and
         on_master_branch(app_defn) and
-        master_pushed_to_remote(app_defn)
+        check_specified_commitish_pushed_to_remote(app_defn, tag)
     )
 
 
 def write_app_yaml(app_defn, gae_app_version=None, preamble=None):
-    # set preamble
+    # set default preamble and GAE app version
     if preamble is None:
         preamble = "\n# *** AUTOMATICALLY GENERATED BY FABRIC ***\n"
+    if gae_app_version is None:
+        gae_app_version = 'default'
 
     # open template
     with open(APP_YAML_TEMPLATE_FILENAME)as app_yaml_template_fd:
@@ -162,7 +218,7 @@ def write_app_yaml(app_defn, gae_app_version=None, preamble=None):
         for key, val in placeholder_replacements:
             placeholder_name = ("$%s_PLACEHOLDER" % key).upper()
             if placeholder_name in yaml:
-                yaml = yaml.replace(placeholder_name, val)
+                yaml = yaml.replace(placeholder_name, unicode(val))
 
     # write app.yaml
     with open(APP_YAML_FILENAME, 'w') as app_yaml_fd:
@@ -172,6 +228,22 @@ def write_app_yaml(app_defn, gae_app_version=None, preamble=None):
 
 def delete_app_yaml():
     os.remove(APP_YAML_FILENAME)
+
+
+def perform_overwrites(app_defn):
+    for src, dest in app_defn.get('overwrite', {}).items():
+        print "Copying %s to %s ..." % (src, dest)
+        shutil.copyfile(src, dest)
+
+
+def write_deployment_version(app_defn, tag, commit):
+    with open('version.json', 'w') as version_json_fd:
+        version_d = {
+            'application': app_defn['application'],
+            'tag': tag,
+            'commit': commit,
+        }
+        json.dump(version_d, version_json_fd)
 
 
 def update():
@@ -206,12 +278,25 @@ def clear_build_dirs():
 
 
 @task
-def check(apps):
+def check(apps, tag='HEAD'):
     """
     Check if it is ok to deploy to apps
     """
     app_defns = get_app_definitions(apps.split(';'))
-    map(ok_to_deploy, app_defns)
+    for app_defn in app_defns:
+        ok_to_deploy(app_defn, tag)
+
+
+@task
+def vacuum_indexes(apps):
+    """
+    Vacuum indexes for apps.
+    """
+    app_defns = get_app_definitions(apps.split(';'))
+    for app_defn in app_defns:
+        local("%(appcfg)s --oauth2 -A %(application)s vacuum_indexes ." % (
+            dict(env.items() + app_defn.items()))
+        )
 
 
 @task
@@ -220,20 +305,27 @@ def deploy(apps, tag='HEAD', version=None):
     Deploy to one or more applications
     (semicolon-separated values or 'all').
     """
-    # if deploying to all, check
-    if apps == 'all':
-        if not confirm("Deploy to ALL apps, excluding sandbox? Are you sure?", default=False):
-            abort("Deploy to all except sandbox unconfirmed")
-
     # get app definitions
     app_defns = get_app_definitions(apps.split(';'))
 
     # before doing anything, check if *all* apps are ok to deploy to
-    map(ok_to_deploy, app_defns)
+    for app_defn in app_defns:
+        ok_to_deploy(app_defn, tag)
 
-    # check tag
-    if not (tag == 'HEAD' or tag in local('git tag -l', capture=True)):
+    # check/rewrite tag
+    available_tags = local('git tag -l', capture=True).splitlines()
+    if tag == 'HEAD':
+        # rewrite HEAD to a tag if available
+        head_tags = local('git tag -l --contains=HEAD', capture=True)
+        if head_tags:
+            tag = head_tags.splitlines()[0]
+        else:
+            warn("HEAD is not tagged with a version.")
+    elif tag not in available_tags:
         abort("Unknown tag '%s'" % tag)
+
+    # get commit hash
+    commit = local('git rev-parse %s' % tag, capture=True)
 
     # decide GAE app version to use
     if version:
@@ -245,24 +337,39 @@ def deploy(apps, tag='HEAD', version=None):
         else:
             gae_app_version = current_branch
 
-    # log that we are deploying
-    print "\n\n** Deploying %s as %s to %s **\n\n" % (tag, gae_app_version, ', '.join(
+    # state options
+    print "\nSelected deployment options:\n"
+    print "Tag:             %s" % tag
+    print "GAE app version: %s" % gae_app_version
+    print "Apps:            %s" % ', '.join(
         app_defn['application'] for app_defn in app_defns
-    ))
+    )
+    print
+
+    # if deploying to all, check
+    if apps == 'all':
+        if not confirm("Deploy to ALL apps, excluding sandbox? Are you sure?", default=False):
+            abort("Deploy to all except sandbox unconfirmed")
 
     # clear old build dirs
     clear_build_dirs()
 
-    # copy dir for deployment
+    # build to deployment, using git archive
     build_dir = mkdtemp(prefix=BUILD_DIR_PREFIX)
-    print "Building to %s" % build_dir
+    print "Building to %s ..." % build_dir
     local("git archive %s | tar -x -C %s" % (tag, build_dir))
+    print "Changing pwd to %s ..." % build_dir
     os.chdir(build_dir)
 
     # deploy to all specified apps
     for app_defn in app_defns:
+        print "\nDeploying to %s...\n" % app_defn['application']
+        print "Writing app.yaml..."
         write_app_yaml(app_defn, gae_app_version=gae_app_version)
-        update()
+        perform_overwrites(app_defn)
+        write_deployment_version(app_defn, tag, commit)
+        print "Starting GAE update..."
+        update()  # call GAE appcfg
         delete_app_yaml()
 
     # output success message
@@ -281,4 +388,25 @@ def write_local_app_yaml():
             "#\n" + 
             "# (edit %s instead)\n\n" % APP_YAML_TEMPLATE_FILENAME
         )
+    )
+
+
+@task
+def dev():
+    " Start development server. "
+    sdk_version_ok()
+    write_local_app_yaml()
+    local(
+        "%s --require_indexes=true --show_mail_body=true ." %
+        os.path.join(env.sdk_path, 'dev_appserver.py')
+    )
+
+
+@task
+def local_shell():
+    " Start a local shell ."
+    sdk_version_ok()
+    print "\nEnter a fake email address and password.\n"
+    local(
+        "%s -s localhost:8080" % os.path.join(env.sdk_path, 'remote_api_shell.py')
     )
